+++ src/app/features/editor/details/components/cp-details/cp-details.component.html
