+++ src/app/features/editor/details/components/info-details/info-details.component.spec.ts
