+++ src/app/features/editor/details/components/information-link-details/information-link-details.component.ts
