+++ src/app/features/editor/details/components/msg-details/msg-details.component.html
