+++ src/app/features/editor/details/services/details.service.spec.ts
