+++ src/app/features/editor/details/services/details.service.ts
@@ -2,10 +2,10 @@ import { Injectable } from '@angular/core';
 import {MatDialog} from "@angular/material/dialog";
 import {ConversationPartner} from "@app/shared/keml/models/core/conversation-partner";
 import {Information, InformationLink, Message} from "@app/shared/keml/models/core/msg-info";
-import {ConversationPartnerDetailsComponent} from "@app/features/editor/features/details/components/cp-details/cp-details.component";
-import {MsgDetailsComponent} from "@app/features/editor/features/details/components/msg-details/msg-details.component";
-import {InfoDetailsComponent} from "@app/features/editor/features/details/components/info-details/info-details.component";
-import {InformationLinkDetailsComponent} from "@app/features/editor/features/details/components/information-link-details/information-link-details.component";
+import {ConversationPartnerDetailsComponent} from "@app/features/editor/details/components/cp-details/cp-details.component";
+import {MsgDetailsComponent} from "@app/features/editor/details/components/msg-details/msg-details.component";
+import {InfoDetailsComponent} from "@app/features/editor/details/components/info-details/info-details.component";
+import {InformationLinkDetailsComponent} from "@app/features/editor/details/components/information-link-details/information-link-details.component";
 
 @Injectable({
   providedIn: 'root'
